modal window for Votings item control buttons
